% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{als}
\alias{als}
\title{ALS Data}
\format{
A data frame with 1822 rows and 371 variables:
\describe{
\item{testset}{Logical variable indicating training (FALSE) or test (TRUE) set}
\item{dFRS}{Response variable: rate of progression of functional rating score}
\item{...}{369 predictor variables based on patient visit measurements}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/ALS.txt}
}
\description{
Data on amyotrophic lateral sclerosis (Lou Gehrig's disease) from Section 17.2.
There are 1822 observations on individuals with ALS. The goal is to predict the
rate of progression dFRS of a functional rating score, using 369 predictors based
on measurements (and derivatives of these) obtained from patient visits.
}
\details{
These data were kindly provided by Lester Mackey and Lilly Fang, who won the
DREAM challenge prediction prize in 2012 (Kuffner et al., 2015). It includes
some additional variables created by them. Their winning entry used Bayesian
trees, not too different from random forests.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Section 17.2.
}
