% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bivnorm}
\alias{bivnorm}
\title{Bivariate Normal Data}
\format{
A data frame with 40 rows and 2 variables:
\describe{
\item{X1}{First variable}
\item{X2}{Second variable}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/bivnorm.csv}
}
\description{
40 points generated from a bivariate normal distribution, with some entries missing.
From Figure 9.3.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Figure 9.3.
}
