% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butterfly}
\alias{butterfly}
\title{Butterfly Species Data}
\format{
A data frame with 24 rows and 2 variables:
\describe{
\item{k}{Number of times a species was trapped}
\item{count}{Number of species seen exactly k times (e.g., 118 species trapped just once, 74 trapped twice each)}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/butterfly.txt}
}
\description{
Number of butterfly species seen a given number of times each in two years of trapping.
From Table 6.2. This is a frequency data frame.
}
\examples{
data(butterfly)
str(butterfly)
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Table 6.2.
}
