% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellinfusion}
\alias{cellinfusion}
\title{Cell Infusion Data}
\format{
A data frame with 25 rows and 4 variables:
\describe{
\item{thrived}{Number of cells that thrived}
\item{N}{Colony size (number of cells)}
\item{ratio}{Ratio of mouse nuclei to human cells (1-5)}
\item{time}{Day of observation (1-5)}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/cellinfusion.txt}
}
\description{
Human cell colonies infused with mouse nuclei in 5 different ratios over 1 to 5 days.
From Table 8.2.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Table 8.2.
}
