% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholesterol}
\alias{cholesterol}
\title{Cholesterol Data}
\format{
A data frame with 164 rows and 2 variables:
\describe{
\item{compliance}{Fraction of intended dose actually taken (standardized)}
\item{cholesterol.decrease}{Decrease in cholesterol level over the course of the experiment}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/cholesterol.txt}
}
\usage{
cholesterol
}
\description{
Cholestyramine, a proposed cholesterol lowering drug, was administered to 164 men
for an average of seven years each. From Figure 20.1.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Figure 20.1.
}
\keyword{datasets}
