% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doseresponse}
\alias{doseresponse}
\title{Dose Response Data}
\format{
A data frame with 11 rows and 2 variables:
\describe{
\item{Dose}{Log dose level (each step is a doubling)}
\item{Proportion}{Proportion of mice that died at that dose}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/doseresponse.txt}
}
\description{
Data from 11 groups of mice (10 each) exposed to drug Xilathon at different doses.
From Figure 8.2.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Figure 8.2.
}
