% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DTI}
\alias{DTI}
\title{DTI Brain Imaging Data}
\format{
A data frame with 15443 rows and 4 variables:
\describe{
\item{x}{Voxel coordinate: back to front}
\item{y}{Voxel coordinate: left to right}
\item{z}{Voxel coordinate: bottom to top}
\item{Zscore}{Z score comparing dyslexic vs normal controls at this voxel}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/DTI.csv}
}
\description{
Diffusion Tensor Imaging (DTI) data comparing 6 dyslexic children with 6 normal
controls, from Figures 15.9 and 15.10. Z scores were computed at 15,443
three-dimensional brain coordinates (voxels).
}
\examples{
data(DTI)
str(DTI)
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Figures 15.9 and 15.10.
}
