% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Insurance Life Table Data}
\format{
A data frame with rows for each age group and 3 variables:
\describe{
\item{age}{Age of policy holders}
\item{n}{Number of policy holders at this age}
\item{y}{Number of deaths at this age}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/insurance.txt}
}
\usage{
insurance
}
\description{
Insurance company life table from Table 9.1. At each age, gives the number of
policy holders and the number of deaths.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Table 9.1.
}
\keyword{datasets}
