% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncog}
\alias{ncog}
\title{NCOG Head and Neck Cancer Data}
\format{
A data frame with survival time information and variables:
\describe{
\item{t}{Time in months until death or censoring}
\item{d}{Death indicator: 1 = death observed, 0 = censored}
\item{arm}{Treatment arm: "A" = Chemotherapy, "B" = Chemotherapy + Radiation}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/ncog.txt}
}
\usage{
ncog
}
\description{
Head and neck cancer survival data from the Northern California Oncology Group
(NCOG), from Section 9.2. Patients were randomized to one of two treatment arms.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Section 9.2.
}
\keyword{datasets}
