% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Lymph Nodes Cancer Data}
\format{
A data frame with 844 rows and 2 variables:
\describe{
\item{n}{Number of lymph nodes removed}
\item{x}{Number of nodes found to be positive (malignant)}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/nodes.txt}
}
\description{
Data on lymph nodes removed from 844 cancer patients, from Figure 6.3.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Figure 6.3.
}
