% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pediatric}
\alias{pediatric}
\title{Pediatric Cancer Survival Data}
\format{
A data frame with 1620 rows and 7 variables:
\describe{
\item{sex}{Sex: 1 = male, 2 = female}
\item{race}{Race: 1 = white, 2 = nonwhite}
\item{age}{Age in years}
\item{entry}{Calendar date of entry in days since July 1, 2001}
\item{far}{Home distance from treatment center in miles}
\item{t}{Survival time in days}
\item{d}{Death indicator: 1 = death observed, 0 = censored}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/pediatric.txt}
}
\description{
Survival data on 1620 children with cancer, from Section 9.4 and Table 9.6.
}
\examples{
data(pediatric)
str(pediatric)
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Section 9.4, Table 9.6.
}
