% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police}
\alias{police}
\title{Police Racial Bias Data}
\format{
A data frame with 2749 rows and 1 variable:
\describe{
\item{z}{Z score measuring potential racial bias}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/police.txt}
}
\usage{
police
}
\description{
Z scores for 2749 New York City police officers, from Figure 15.7. A large
value suggests racial bias in policing behavior.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Figure 15.7.
}
\keyword{datasets}
