% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostz}
\alias{prostz}
\title{Prostate Cancer Z-values}
\format{
A data frame with 6033 rows and 1 variable:
\describe{
\item{z}{Z-value for each gene comparing cancer vs control expression}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/prostz.txt}
}
\description{
Vector of 6033 z-values comparing gene expression between prostate cancer
patients and controls, as pictured in Figure 3.4. These were computed as
described on page 272.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Section 3.3, Figure 3.4.
}
