% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_kernel}
\alias{protein_kernel}
\title{Protein Classification Kernel Matrix}
\format{
A 1708 x 1708 numeric matrix of inner products.
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/protein_kernel.txt}
}
\usage{
protein_kernel
}
\description{
A 1708 x 1708 inner-product (kernel) matrix derived from 1708 human proteins,
from Section 19.6. Each protein is represented using a bag of N-grams (N=4)
of amino acids, and the kernel was computed using a string kernel (Leslie et al., 2003).
}
\details{
The corresponding labels are in the \code{protein_label} dataset (not included).
}
\references{
Leslie, C., Eskin, E., Cohen, A., Weston, J. and Noble, W.S. (2003). Mismatch
String Kernels for SVM Protein Classification. \emph{Advances in NIPS 15}.

Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Section 19.6.
}
\keyword{datasets}
