% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{student_score}
\alias{student_score}
\title{Student Score Data}
\format{
A data frame with 22 rows and 5 variables:
\describe{
\item{mech}{Mechanics exam score}
\item{vecs}{Vectors exam score}
\item{alg}{Algebra exam score}
\item{analy}{Analysis exam score}
\item{stat}{Statistics exam score}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/student_score.txt}
}
\description{
Test scores for 22 students on 5 different exams, from Tables 3.1 and 10.1.
}
\examples{
data(student_score)
str(student_score)
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Tables 3.1 and 10.1.
}
