% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vasoconstriction}
\alias{vasoconstriction}
\title{Vasoconstriction Data}
\format{
A data frame with 39 rows and 2 variables:
\describe{
\item{volume}{Volume measurement}
\item{constriction}{Logical: TRUE if constriction occurred, FALSE otherwise}
}
}
\source{
\url{https://hastie.su.domains/CASI_files/DATA/vasoconstriction.txt}
}
\description{
Data on vasoconstriction (lung constriction) response, from Table 13.2.
}
\references{
Efron, B. and Hastie, T. (2016). \emph{Computer Age Statistical Inference}.
Cambridge University Press, Table 13.2.
}
